pipeline-python-mongo-mysql/requirements.txt
requests
pymongo
python-dotenv
pandas
mysql-connector-python
dnspython